whisper_server/requirements.txt
openai-whisper
fastapi
uvicorn
python-multipart
torch
